Aya b

+++ database/migrations/2023_01_27_163415_create_reglements_table.php
@@ -18,7 +18,7 @@ class CreateReglementsTable extends Migration
 
             $table->date('dateR');
             $table->string('methode');
-            $table->float('montant');
+            $table->double('montant');
             $table->integer('numero_cheque')->nullable();
             $table->unsignedBigInteger('id_facture')->nullable();
             $table->foreign('id_facture')->references('id')->on('factures');
